+++ thejoker/tests/test_samples.py
@@ -8,7 +8,7 @@ import pytest
 
 # Project
 from ..samples import JokerSamples
-from ..samples_helpers import _custom_tbl_dtype_compare, is_P_unimodal
+from ..samples_helpers import _custom_tbl_dtype_compare
 
 
 def test_joker_samples(tmpdir):